Evaluating Urban Trees in Los Angeles
This tool uses imagery, census block boundaries, and tree canopy metadata (in the form of a csv) in Los Angeles to identify census blocks in most need of increased tree canopy. It evaluates census blocks' existing tree canopy in 2016 (percent of census block area with tree canopy), percent of area identified in having potential for additional tree canopy as of 2016, and an updated percent of existing tree canopy in 2022.